refactor: centralize experience schemas
