+++ src/entity/user/experiences/UserCertificationExperience.ts
@@ -1,8 +1,7 @@
 import { UserExperience } from './UserExperience';
 import { ChildEntity, Column, JoinColumn, ManyToOne } from 'typeorm';
 import type { Company } from '../../Company';
-import { UserExperienceType, baseUserExperienceSchema } from './types';
-import { z } from 'zod';
+import { UserExperienceType } from './types';
 
 @ChildEntity(UserExperienceType.Certification)
 export class UserCertificationExperience extends UserExperience {
@@ -23,17 +22,3 @@ export class UserCertificationExperience extends UserExperience {
   @Column({ type: 'text', nullable: true })
   credentialUrl: string;
 }
-
-// Zod schema for UserCertificationExperience
-export const userCertificationExperienceSchema =
-  baseUserExperienceSchema.extend({
-    type: z.literal(UserExperienceType.Certification),
-    courseNumber: z.string().nullable().optional(),
-    companyId: z.string().uuid(),
-    credentialId: z.string().nullable().optional(),
-    credentialUrl: z
-      .string()
-      .url('Credential URL must be a valid URL')
-      .nullable()
-      .optional(),
-  });
