+++ src/entity/user/experiences/UserEducationExperience.ts
@@ -1,8 +1,7 @@
 import { ChildEntity, Column, JoinColumn, ManyToOne } from 'typeorm';
 import { UserExperience } from './UserExperience';
 import type { Company } from '../../Company';
-import { UserExperienceType, baseUserExperienceSchema } from './types';
-import { z } from 'zod';
+import { UserExperienceType } from './types';
 
 @ChildEntity(UserExperienceType.Education)
 export class UserEducationExperience extends UserExperience {
@@ -23,12 +22,3 @@ export class UserEducationExperience extends UserExperience {
   @Column({ type: 'text', nullable: true })
   extracurriculars: string;
 }
-
-// Zod schema for UserEducationExperience
-export const userEducationExperienceSchema = baseUserExperienceSchema.extend({
-  type: z.literal(UserExperienceType.Education),
-  schoolId: z.string().uuid(),
-  fieldOfStudy: z.string().min(1, 'Field of study is required'),
-  grade: z.string().nullable().optional(),
-  extracurriculars: z.string().nullable().optional(),
-});
