+++ src/entity/user/experiences/types.ts
@@ -1,5 +1,3 @@
-import { z } from 'zod';
-
 export enum UserExperienceType {
   Work = 'work',
   Education = 'education',
@@ -43,21 +41,3 @@ export type ProjectLink = {
   type: ProjectLinkType;
   url: string;
 };
-
-// Zod schema for ProjectLink
-export const projectLinkSchema = z.object({
-  type: z.nativeEnum(ProjectLinkType),
-  url: z.string().url('URL must be a valid URL'),
-});
-
-// Base Zod schema for UserExperience
-export const baseUserExperienceSchema = z.object({
-  id: z.string().uuid().optional(), // Optional for creation
-  userId: z.string().uuid(),
-  title: z.string().min(1, 'Title is required'),
-  description: z.string().default(''),
-  startDate: z.date(),
-  endDate: z.date().nullable().optional(),
-  status: z.nativeEnum(ExperienceStatus).default(ExperienceStatus.Draft),
-  flags: z.record(z.string(), z.unknown()).default({}),
-});
